+++ nac_collector/controller/__init__.py
@@ -1,0 +1,1 @@
+# Controller-based Cisco client implementations
