+++ nac_collector/device/__init__.py
@@ -1,0 +1,1 @@
+# Device-based Cisco client implementations
