+++ tests/unit/ise/test_authentication.py
@@ -2,7 +2,7 @@ from unittest.mock import Mock
 
 import pytest
 
-from nac_collector.cisco_client_ise import CiscoClientISE
+from nac_collector.controller.ise import CiscoClientISE
 
 pytestmark = pytest.mark.unit
