+++ tests/unit/ise/test_initialization.py
@@ -1,6 +1,6 @@
 import pytest
 
-from nac_collector.cisco_client_ise import CiscoClientISE
+from nac_collector.controller.ise import CiscoClientISE
 
 pytestmark = pytest.mark.unit
